Authentication&Authorization

+++ backend/controllers/feed.js
@@ -2,10 +2,30 @@ const { validationResult } = require('express-validator');
 const fs = require('fs');
 const path = require('path');
 const Post = require('../models/post');
+const User = require('../models/user');
 exports.getPosts = (req, res, next) => {
+  const currentPage = req.query.page || 1;
+  const postsPerPage = 2;
+  let totalPosts;
   Post.find()
+    .countDocuments()
+    .then((count) => {
+      totalPosts = count;
+      return Post.find()
+        .skip((currentPage - 1) * postsPerPage)
+        .limit(postsPerPage)
+        .populate('creator', 'name');
+      //populate will take 2 params.
+      //first parameter will be the new property that will be added to each element found
+      //second parameter will be the value of the property that will be fetched from the linked Model [ here name will be
+      // fetched from the User model because it is referenced in the model definition]
+    })
     .then((posts) => {
-      res.status(200).json({ message: 'Fetched Posts.', posts: posts });
+      return res.status(200).json({
+        message: 'Fetched Posts.',
+        posts: posts,
+        totalItems: totalPosts,
+      });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -33,18 +53,31 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace('\\', '/');
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title,
     content,
     imageUrl: imageUrl,
-    creator: { name: 'surya' },
+    creator: req.userId,
+    //when authenticated, we added userId property to the req object
   });
   post
     .save()
     .then((result) => {
+      //before sending the response, add the post to the user.
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      //adding the post to the user.posts -> mapping each post to an user
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       return res.status(201).json({
         message: 'successfully created',
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -60,6 +93,7 @@ exports.createPost = (req, res, next) => {
 exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
+    .populate('creator', 'name')
     .then((post) => {
       //if post is empty
       if (!post) {
@@ -110,6 +144,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+
       //delete the previous file if it's changed
       if (imageUrl !== post.imageUrl) {
         deleteFile(post.imageUrl);
@@ -131,6 +171,45 @@ exports.updatePost = (req, res, next) => {
         err.statusCode = 500;
       }
       next(err);
+    });
+};
+
+exports.deletePost = (req, res, next) => {
+  const postId = req.params.postId;
+  Post.findById(postId)
+    .then((post) => {
+      if (!post) {
+        const error = new Error('Post Not Found.');
+        error.statusCode = 404;
+        throw error;
+      }
+      //check if the loggedin user is same as the creator of the post
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+      //delete file
+      deleteFile(post.imageUrl);
+      return Post.findByIdAndRemove(postId);
+    })
+    .then((result) => {
+      //after deleting the post, delete the post reference from the user
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      //mongoose provides pull method which we can use to delete the element based on the id.
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
+      return res.status(200).json({ message: 'Deleted successfully.' });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+        next(err);
+      }
     });
 };
